Successfully separated functions within customer.java
