+++ bank/nbproject/private/private.xml
@@ -3,8 +3,8 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/DeLaCruz/Desktop/bank/src/bank/Account.java</file>
-            <file>file:/C:/Users/DeLaCruz/Desktop/bank/src/bank/Customer.java</file>
+            <file>file:/C:/Users/DeLaCruz/Desktop/Repositories/bank-app/bank/src/bank/Customer.java</file>
+            <file>file:/C:/Users/DeLaCruz/Desktop/Repositories/bank-app/bank/src/bank/Account.java</file>
         </group>
     </open-files>
 </project-private>
